chore : entity id를 Long 타입으로 변경, 어떤 entity인지 명확히 드러나도록 필드명 변경

+++ Kakao-Backend-Second-DTO/src/main/java/com/example/kakaoshop/product/domain/model/ProductOption.java
@@ -1,6 +1,5 @@
 package com.example.kakaoshop.product.domain.model;
 
-import com.example.kakaoshop.product.entity.ProductEntity;
 import lombok.*;
 
 
@@ -11,7 +10,7 @@ import lombok.*;
 @EqualsAndHashCode
 @ToString
 public class ProductOption {
-    private int id;
+    private Long id;
     private Product product;
     private String optionName;
     private int price;
